Moved opengl func in a separate file
