+++ scrutiny/gui/main_window.py
@@ -18,7 +18,7 @@ import PySide6QtAds  as QtAds   # type: ignore
 from scrutiny import tools
 from scrutiny.gui.app_settings import app_settings
 from scrutiny.gui.tools.invoker import InvokeQueued
-from scrutiny.gui.tools import prepare_for_opengl
+from scrutiny.gui.tools.opengl import prepare_for_opengl
 
 from scrutiny.gui.dialogs.about_dialog import AboutDialog
 from scrutiny.gui.widgets.component_sidebar import ComponentSidebar
@@ -159,7 +159,7 @@ class MainWindow(QMainWindow):
         try:
             widget = component_class(self, name, self._watchable_registry, self._server_manager)
             if app_settings().opengl_enabled:
-                prepare_for_opengl(widget)  # On every widget. Flaoting widget creates a new window -> Must be done on each window
+                prepare_for_opengl(widget)  # On every widget. Flaating widget creates a new window -> Must be done on each window
         except Exception as e:
             tools.log_exception(self._logger, e, f"Failed to create a dashboard component of type {component_class.__name__}")    
             return
